Fixed a bug where in certain AIFF files, the ID3 tag wasn't detected correctly, leading to a CorruptFileException.
The problem was that, to find the ID3 tag, the reader is simply scanning the file for the byte sequence ID3. This can occur naturally within, say, the SSND chunk. The fix included here is to make sure that the ID3 sequence anticipated to be the chunk header is actually followed 8 bytes later by the expected ID3 sequence denoting the actual ID3 tag.